Zustand to the rescue?
Attempted to solve the animation control issue by creating a Provider component that communicates with a Zustand store to manage the ready state of the tsparticles engine.
This works better, but there's a bug that prevents the buttons from working when navigating away from the page and then back to it. There may be a miscommunication somewhere that needs to be fixed.

In this current commit, the Pause button handles toggling between play and pause. The functionality stops working the moment I introduce a state variable in the jsx for condition rendering...
The "particlesLoaded" function seems to re-render the component for every animation frame of the <canvas>, but this is only made clear the moment you try to use any state near it. If you just console log the container, it only logs once (or twice). But if you try to setParticleContainer with useState, the particleContainer seems to be set in an infinite loop.
If you use an unrelated "isPlaying" variable and try to render a play OR pause button depending on the state, then the animation fails to comply. You can see a visible flash in the animation, but it instantly re-renders from its current position and the pause command is ignored.

• Did some file relocation/renaming