Carousels Complete
Solved the second particles canvas re-rendering issue by assigning the <Particles /> to a memoized const. I can now render a play or pause button depending on an 'isPlaying' state. I tried to hook it up to the Zustand store to make the last chosen animation state persist across page navigations, but this didn't work. That's ok though. I'm happy with where it is now.

Quickly knocked out the <AutoScrollCarousel />.

• Updated README
• Updated <ParticlesProvider>
• Added a mobile friendly navigation to the <Header /> so the browser window can shrink further before forcing a horizontal scrollbar.
• Removed some public assets and the <UserButton />
• Adjusted the fireworksParticles to look a bit nicer.